+++ actions/snowflake-data/CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.0.3] - 2025-09-23
 
 ### Changed
-- **IMPROVED**: Updated `get_tables_info` action to return a more informative error message if no tables are found
+- `get_tables_info` action returns a more informative error message if no tables are found
+- `get_tables_info` action includes views in the results
+- `get_tables_info` action now accepts a `table_filter` parameter to limit the results to a specific table or pattern of tables
 
 ## [0.0.2] - 2025-09-19